requirements.txt
django
djangorestframework
dj-database-url
python-decouple
dj-static